requirements.txt
pandas
numpy
scikit-learn
requests
plotly
streamlit
umap-learn

